refactor(cloud)!: Use 'quota' as default vs. 'quotas'
