+++ internal/cli/kraft/cloud/cloud.go
@@ -18,7 +18,7 @@ import (
 	"kraftkit.sh/internal/cli/kraft/cloud/image"
 	"kraftkit.sh/internal/cli/kraft/cloud/instance"
 	"kraftkit.sh/internal/cli/kraft/cloud/metro"
-	"kraftkit.sh/internal/cli/kraft/cloud/quotas"
+	"kraftkit.sh/internal/cli/kraft/cloud/quota"
 	"kraftkit.sh/internal/cli/kraft/cloud/scale"
 	"kraftkit.sh/internal/cli/kraft/cloud/service"
 	"kraftkit.sh/internal/cli/kraft/cloud/tunnel"
@@ -86,7 +86,7 @@ func NewCmd() *cobra.Command {
 	}
 
 	cmd.AddCommand(deploy.NewCmd())
-	cmd.AddCommand(quotas.NewCmd())
+	cmd.AddCommand(quota.NewCmd())
 	cmd.AddCommand(tunnel.NewCmd())
 
 	cmd.AddGroup(&cobra.Group{ID: "kraftcloud-image", Title: "IMAGE COMMANDS"})
